third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-start-before-get-set.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<style>
.gridItemWithPositiveInteger {
    -webkit-grid-start: 10;
    -webkit-grid-before: 15;
}
.gridItemWithNegativeInteger {
    -webkit-grid-start: -10;
    -webkit-grid-before: -15;
}
.gridItemWithBeforeSpan {
    -webkit-grid-start: span 2;
    -webkit-grid-before: span 8;
}
.gridItemWithAfterSpan {
    -webkit-grid-start: 2 span;
    -webkit-grid-before: 8 span;
}
.gridItemWithOnlySpan {
    -webkit-grid-start: span;
    -webkit-grid-before: span;
}
.gridItemWithAuto {
    -webkit-grid-start: auto;
    -webkit-grid-before: auto;
}
.gridItemWithString {
    -webkit-grid-start: "first";
    -webkit-grid-before: "last";
}
.gridItemWithSpanString {
    -webkit-grid-start: "first" span;
    -webkit-grid-before: span "last";
}
.gridItemWithSpanNumberString {
    -webkit-grid-start: 2 "first" span;
    -webkit-grid-before: "last" 3 span;
}
</style>
<script src="resources/grid-item-column-row-parsing-utils.js"></script>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<!-- The first has no properties set on it. -->
<div id="gridElement"></div>
<div class="gridItemWithPositiveInteger" id="gridItemWithPositiveInteger"></div>
<div class="gridItemWithNegativeInteger" id="gridItemWithNegativeInteger"></div>
<div class="gridItemWithBeforeSpan" id="gridItemWithBeforeSpan"></div>
<div class="gridItemWithAfterSpan" id="gridItemWithAfterSpan"></div>
<div class="gridItemWithOnlySpan" id="gridItemWithOnlySpan"></div>
<div class="gridItemWithAuto" id="gridItemWithAutoElement"></div>
<div class="gridItemWithString" id="gridItemWithStringElement"></div>
<div class="gridItemWithSpanString" id="gridItemWithSpanStringElement"></div>
<div class="gridItemWithSpanNumberString" id="gridItemWithSpanNumberStringElement"></div>
<script>
    description('Test that setting and getting grid-start and grid-before works as expected');

    debug("Test getting -webkit-grid-start and -webkit-grid-before set through CSS");
    var gridElement = document.getElementById("gridElement");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-start')", "'auto'");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-before')", "'auto'");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");

    testColumnRowCSSParsing("gridItemWithPositiveInteger", "10 / auto", "15 / auto");
    testColumnRowCSSParsing("gridItemWithNegativeInteger", "-10 / auto", "-15 / auto");
    testColumnRowCSSParsing("gridItemWithBeforeSpan", "span 2 / auto", "span 8 / auto");
    testColumnRowCSSParsing("gridItemWithAfterSpan", "span 2 / auto", "span 8 / auto");
    testColumnRowCSSParsing("gridItemWithOnlySpan", "span 1 / auto", "span 1 / auto");
    testColumnRowCSSParsing("gridItemWithAutoElement", "auto / auto", "auto / auto");
    testColumnRowCSSParsing("gridItemWithStringElement", "1 first / auto", "1 last / auto");
    testColumnRowCSSParsing("gridItemWithSpanStringElement", "span 1 first / auto", "span 1 last / auto");
    testColumnRowCSSParsing("gridItemWithSpanNumberStringElement", "span 2 first / auto", "span 3 last / auto");

    debug("");
    debug("Test the initial value");
    var element = document.createElement("div");
    document.body.appendChild(element);
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-start')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-before')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");

    debug("");
    debug("Test getting and setting grid-start and grid-before through JS");
    testStartBeforeJSParsing("18", "66");
    testStartBeforeJSParsing("-55", "-40");
    testStartBeforeJSParsing("'nav'", "'last'", "1 nav", "1 last");
    testStartBeforeJSParsing("span 3", "span 20");
    testStartBeforeJSParsing("span 'nav'", "span 'last'", "span 1 nav", "span 1 last");
    testStartBeforeJSParsing("auto", "auto");

    debug("");
    debug("Test setting grid-start and grid-before to 'inherit' through JS");
    testStartBeforeInheritJSParsing("inherit", "18");
    testStartBeforeInheritJSParsing("2", "inherit");
    testStartBeforeInheritJSParsing("inherit", "inherit");

    debug("");
    debug("Test setting grid-start and grid-before to 'initial' through JS");
    testStartBeforeInitialJSParsing();

    debug("");
    debug("Test setting grid-start and grid-before back to 'auto' through JS");
    element.style.webkitGridStart = "18";
    element.style.webkitGridBefore = "66";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-start')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'18 / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-before')", "'66'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'66 / auto'");
    element.style.webkitGridStart = "auto";
    element.style.webkitGridBefore = "auto";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-start')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto / auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-before')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto / auto'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
